added river template
I figured we should keep showing templates for interesting data types like rivers/trees, so that students (hopefully) get in the habit of starting from a template when designing functions. I made the template for the River data type specifically, since that's how we're covering trees here, but it could also be a general BinaryTree template.